// ClientApp/src/app/models/payment.ts
export class Payment {
  Amount: number | undefined;
  UserID: number | undefined;
  Email: string = '';
  CourseID: number | undefined;
  Title: string = '';
  Nit: number | undefined;
  RazonSocial: string = '';
  Price: number | undefined;
  IsOnline: boolean = false;
  IsOppened: boolean = false;
}